refactor: move guard to guard folder
