--- src/auth/auth.guard.spec.ts
@@ -1,7 +1,0 @@
-import { AuthGuard } from './auth.guard';
-
-describe('AuthGuard', () => {
-  it('should be defined', () => {
-    expect(new AuthGuard()).toBeDefined();
-  });
-});
