+++ src/auth/guard/auth.guard.ts
@@ -9,7 +9,7 @@ import { JwtService } from '@nestjs/jwt';
 import { Request } from 'express';
 import { REFRESH_ROUTE } from 'src/constants/constants';
 import { IS_PUBLIC_KEY } from 'src/decorators/public.decorator';
-import { RefreshDto } from './dto/refresh.dto';
+import { RefreshDto } from '../dto/refresh.dto';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
